% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-convert.R
\name{geo_convert}
\alias{geo_convert}
\title{Convert a geometry to a different data structure}
\usage{
geo_convert(x, to)
}
\arguments{
\item{x}{A geometry-like object}

\item{to}{A prototype created by \code{\link[=geo_ptype]{geo_ptype()}}}
}
\value{
A geometry vector, in the format defined by \code{to}.
}
\description{
Convert a geometry to a different data structure
}
\examples{
geo_convert(geo_wkt("POINT (20 10)"), geo_wkb())
geo_convert(geo_wkt("POINT (20 10)"), geo_coord())

}
