% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-coord-linestring.R, R/geo-coord-point.R,
%   R/geo-coord-polygon.R
\name{geo_coord_linestring}
\alias{geo_coord_linestring}
\alias{geo_coord_multilinestring}
\alias{geo_coord_point}
\alias{geo_coord_multipoint}
\alias{geo_coord_polygon}
\alias{geo_coord_multipolygon}
\title{Coordinate-long representations of geometries}
\usage{
geo_coord_linestring(xy, feature = 1L)

geo_coord_multilinestring(xy, feature = 1L, part = 1L)

geo_coord_point(xy, feature = seq_len(vec_size(xy)))

geo_coord_multipoint(xy, feature = seq_len(vec_size(xy)))

geo_coord_polygon(xy, feature = 1L, piece = 1L)

geo_coord_multipolygon(xy, feature = 1L, piece = 1L, part = 1L)
}
\arguments{
\item{xy}{A \code{\link[=geo_xy]{geo_xy()}} (or something castable to \code{\link[=geo_xy]{geo_xy()}}) of coordinates}

\item{feature}{An integer ID that whose unique values can
be used to identify and/or separate features.}

\item{part}{An ID used to separate parts within a feature of a multi* geometry}

\item{piece}{An ID used to separate rings within polygons. The
smallest \code{piece} value within a \code{feature}/\code{part} combination
is assumed to be the outer ring.}
}
\value{
A vctr
}
\description{
R's graphics use coordinate-long representations of geometries (i.e.,
one row per coordinate). Additionally, these can be more easily manipulated
than some other representations of geometry. \strong{Important}: order matters
when passing a \verb{geo_coord_*()} to other functions, which assume
that you use sequential \code{feature}, \code{piece}, and \code{part} values starting at 1.
Helpers to construct this particular representation of geometry from
a tibble are in the works.
}
\examples{
geo_coord_point(geo_xy(30, 10))
geo_coord_multipoint(geo_xy(1:4, 1:4), feature = c(1, 1, 2, 2))

}
