% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-binary-operators.R
\name{geo_intersection}
\alias{geo_intersection}
\title{Intersect two geometry vectors}
\usage{
geo_intersection(x, y, to = geo_ptype(x))
}
\arguments{
\item{x}{A geometry-like object}

\item{y}{A geometry-like object}

\item{to}{A prototype created by \code{\link[=geo_ptype]{geo_ptype()}}}
}
\value{
A geometry-like object, in the format defined by \code{to}.
}
\description{
Intersect two geometry vectors
}
\examples{
poly1 <- geo_wkt("POLYGON ((0 0, 0 10, 10 10, 10 0, 0 0))")
poly2 <- geo_wkt("POLYGON ((5 5, 5 15, 15 15, 15 5, 5 5))")

geo_plot(c(poly1, poly2))
geo_plot_add(geo_intersection(poly1, poly2), col = "grey90")

}
