% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-unary-vector-operators.R
\name{geo_is_empty}
\alias{geo_is_empty}
\title{Check for empty geometries}
\usage{
geo_is_empty(x)
}
\arguments{
\item{x}{A geometry-like object}
}
\value{
A logical vector
}
\description{
Check for empty geometries
}
\examples{
geo_is_empty(geo_wkt("POINT EMPTY"))
geo_is_empty(geo_wkt("POINT (30 10)"))

}
