% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-ptype.R
\name{geo_ptype}
\alias{geo_ptype}
\alias{geo_ptype.default}
\alias{geo_restore}
\alias{geo_restore.default}
\alias{geo_size}
\alias{geo_size.default}
\title{Geometry-specific coersion}
\usage{
geo_ptype(x)

\method{geo_ptype}{default}(x)

geo_restore(to, x)

\method{geo_restore}{default}(to, x)

geo_size(x)

\method{geo_size}{default}(x)
}
\arguments{
\item{x}{A geometry-like object}

\item{to}{A prototype created by \code{\link[=geo_ptype]{geo_ptype()}}}
}
\description{
Similar to \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[vctrs:vec_ptype]{vctrs::vec_ptype()}},
except that some important geometry data structures are
not strictly vctrs, in the sense that \code{\link[vctrs:vec_size]{vctrs::vec_size()}}
may be different or not applicable (e.g., for a \code{\link[=geo_coord_linestring]{geo_coord_linestring()}}).
These functions help geometry operators accept a number of different
input and output formats. In particular, \code{\link[=geo_restore]{geo_restore()}} provides
any conversion that may be necessary between the C++ data structure
and the R data  structure.
}
