% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-rect.R
\name{geo_rect}
\alias{geo_rect}
\title{Rectangles}
\usage{
geo_rect(xmin = double(), ymin = double(), xmax = double(), ymax = double())
}
\arguments{
\item{xmin, ymin, xmax, ymax}{Border values, recycled to a common
length using \code{\link[vctrs:vec_recycle_common]{vctrs::vec_recycle_common()}}.}
}
\value{
A \code{\link[=new_geo_rect]{new_geo_rect()}}
}
\description{
Rectangles
}
\examples{
geo_rect(xmin = 0:5, ymin = 0:5, xmax = 2:7, ymax = 2:7)

}
