% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-segment.R
\name{geo_segment}
\alias{geo_segment}
\title{Line segments}
\usage{
geo_segment(start = geo_xy(), end = geo_xy())
}
\arguments{
\item{start, end}{\code{\link[=geo_xy]{geo_xy()}}s for the start and end
of the segment, respectively.}
}
\value{
A \code{\link[=new_geo_segment]{new_geo_segment()}}
}
\description{
Line segments
}
\examples{
geo_segment(start = geo_xy(), end = geo_xy())

}
