% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-tbl-linestring.R, R/geo-tbl-point.R,
%   R/geo-tbl-polygon.R
\name{geo_tbl_linestring}
\alias{geo_tbl_linestring}
\alias{geo_tbl_multilinestring}
\alias{geo_tbl_point}
\alias{geo_tbl_multipoint}
\alias{geo_tbl_polygon}
\alias{geo_tbl_multipolygon}
\title{Coordinate-long representations of geometries}
\usage{
geo_tbl_linestring(xy, feature)

geo_tbl_multilinestring(xy, feature, part = 1L)

geo_tbl_point(xy, feature = seq_len(vec_size(xy)))

geo_tbl_multipoint(xy, feature)

geo_tbl_polygon(xy, feature, piece = 1L)

geo_tbl_multipolygon(xy, feature = 1L, piece = 1L, part = 1L)
}
\arguments{
\item{xy}{A \code{\link[=geo_xy]{geo_xy()}} (or something castable to \code{\link[=geo_xy]{geo_xy()}}) of coordinates}

\item{feature}{An integer ID that whose unique values can
be used to identify and/or separate features}

\item{part}{An ID used to separate parts within a feature of a multi* geometry}

\item{piece}{An ID used to separate rings within polygons. The
first piece in the row order is interpreted as the outer ring.}
}
\value{
A vctr
}
\description{
Coordinate-long representations of geometries
}
\examples{
geo_tbl_point(geo_xy(30, 10))
geo_tbl_multipoint(geo_xy(1:4, 1:4), feature = c(1, 1, 2, 2))

}
