% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-wkb.R
\name{geo_wkb}
\alias{geo_wkb}
\title{Create and validate well-known binary}
\usage{
geo_wkb(x = list())
}
\arguments{
\item{x}{A \code{\link[=list]{list()}} of \code{\link[=raw]{raw()}} objects, each of which
represent well-known binary}
}
\value{
A \code{\link[=new_geo_wkb]{new_geo_wkb()}}
}
\description{
Create and validate well-known binary
}
\examples{
# POINT (30 10) in WKB
wkb <- as.raw(
  c(
    0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
    0x00, 0x00, 0x00, 0x3e, 0x40, 0x00, 0x00, 0x00,
    0x00, 0x00, 0x00, 0x24, 0x40
  )
)
geo_wkb(list(wkb))

}
