% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-xy.R
\name{geo_xy}
\alias{geo_xy}
\title{Create a coordinate vector}
\usage{
geo_xy(x = double(), y = double())
}
\arguments{
\item{x, y}{x and y coordinates}
}
\value{
A \code{\link[=new_geo_xy]{new_geo_xy()}}
}
\description{
Create a coordinate vector
}
\examples{
geo_xy(0:5, 1:6)

}
