% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-package.R
\docType{package}
\name{geom-package}
\alias{geom}
\alias{geom-package}
\title{geom: Geometry vectors}
\description{
Provides data structures to convert various
  representations of geometry.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
