% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-package.R
\docType{package}
\name{geos-package}
\alias{geos}
\alias{geos-package}
\title{geos: Open Source Geometry Engine ('GEOS') R API}
\description{
Provides an R API to the Open Source Geometry Engine ('GEOS') library (<https://trac.osgeo.org/geos/>) and a vector format with which to efficiently store 'GEOS' geometries. High-performance functions to extract information from, calculate relationships between, and transform geometries are provided. Finally, facilities to import and export geometry vectors to other spatial formats are provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paleolimbot.github.io/geos/}
  \item \url{https://github.com/paleolimbot/geos/}
  \item Report bugs at \url{https://github.com/paleolimbot/geos/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

Authors:
\itemize{
  \item Edzer Pebesma \email{edzer.pebesma@uni-muenster.de} (\href{https://orcid.org/0000-0001-8049-7069}{ORCID})
}

}
\keyword{internal}
