% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-misc.R
\name{geos_area}
\alias{geos_area}
\alias{geos_length}
\alias{geos_distance}
\title{Area, length, and distance}
\usage{
geos_area(x)

geos_length(x)

geos_distance(x, y)
}
\arguments{
\item{x}{A geometry-like object}

\item{y}{A geometry-like object. \code{x} and \code{y} of length 1
is recycled to the length of the other (like \code{\link[tibble:tibble]{tibble::tibble()}}).}
}
\value{
\itemize{
\item \code{\link[=geos_area]{geos_area()}} computes areas for polygons, or returns 0 otherwise.
\item \code{\link[=geos_length]{geos_length()}} computes the length of the boundary for polygons, or the length
of the line for linestrings.
\item \code{\link[=geos_distance]{geos_distance()}} returns the smallest possible distance between the two
geometries.
}
}
\description{
Area, length, and distance
}
\examples{
geos_area(geo_wkt("POLYGON ((0 0, 10 0, 0 10, 0 0))"))
geos_length(geo_wkt("POLYGON ((0 0, 10 0, 0 10, 0 0))"))
geos_distance(
  geo_wkt("POLYGON ((0 0, 10 0, 0 10, 0 0))"),
  geo_wkt("POINT (10 10)")
)

}
