% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-make.R
\name{geos_create_rectangle}
\alias{geos_create_rectangle}
\title{Create rectangles from bounds}
\usage{
geos_create_rectangle(xmin, ymin, xmax, ymax, crs = NULL)
}
\arguments{
\item{xmin}{Left bound of envelope}

\item{ymin}{Lower bound of envelope}

\item{xmax}{Right bound of envelope}

\item{ymax}{Upper bound of envelope}

\item{crs}{An object that can be interpreted as a CRS. See \code{\link[wk:wk_crs]{wk::wk_crs()}}.}
}
\value{
A \code{\link[=geos_geometry]{geos_geometry()}} consisting of a polygon
}
\description{
Create rectangles from bounds
}
