% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-strtree.R
\name{geos_disjoint_matrix}
\alias{geos_disjoint_matrix}
\alias{geos_touches_matrix}
\alias{geos_intersects_matrix}
\alias{geos_crosses_matrix}
\alias{geos_within_matrix}
\alias{geos_contains_matrix}
\alias{geos_overlaps_matrix}
\alias{geos_equals_matrix}
\alias{geos_covers_matrix}
\alias{geos_covered_by_matrix}
\title{Matrix predicates}
\usage{
geos_disjoint_matrix(geom1, geom2)

geos_touches_matrix(geom1, geom2)

geos_intersects_matrix(geom1, geom2)

geos_crosses_matrix(geom1, geom2)

geos_within_matrix(geom1, geom2)

geos_contains_matrix(geom1, geom2)

geos_overlaps_matrix(geom1, geom2)

geos_equals_matrix(geom1, geom2)

geos_covers_matrix(geom1, geom2)

geos_covered_by_matrix(geom1, geom2)
}
\value{
A \code{list()} of integer vectors containing the indices of \code{tree}
for which the predicate would return TRUE.
}
\description{
Matrix predicates
}
