% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-binary-operators.R
\name{geos_intersection}
\alias{geos_intersection}
\alias{geos_difference}
\alias{geos_sym_difference}
\alias{geos_union}
\alias{geos_unary_union}
\alias{geos_coverage_union}
\alias{geos_clip_by_rect}
\title{Intersect two geometry vectors}
\usage{
geos_intersection(x, y, to = geo_ptype(x))

geos_difference(x, y, to = geo_ptype(x))

geos_sym_difference(x, y, to = geo_ptype(x))

geos_union(x, y, to = geo_ptype(x))

geos_unary_union(x, to = geo_ptype(x))

geos_coverage_union(x, to = geo_ptype(x))

geos_clip_by_rect(x, rect, to = geo_ptype(x))
}
\arguments{
\item{x}{A geometry-like object}

\item{y}{A geometry-like object. \code{x} and \code{y} of length 1
is recycled to the length of the other (like \code{\link[tibble:tibble]{tibble::tibble()}}).}

\item{to}{A prototype created by \code{\link[=geo_ptype]{geo_ptype()}}}

\item{rect}{A \code{\link[=geo_rect]{geo_rect()}} used for the non-robust clipping
algorithm.}
}
\value{
A geometry-like object, in the format defined by \code{to}.
}
\description{
Intersect two geometry vectors
}
\examples{
poly1 <- geo_wkt("POLYGON ((0 0, 0 10, 10 10, 10 0, 0 0))")
poly2 <- geo_wkt("POLYGON ((5 5, 5 15, 15 15, 15 5, 5 5))")

geo_plot(c(poly1, poly2))
geo_plot_add(geos_intersection(poly1, poly2), col = "grey90")

}
