% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-unary-vector-operators.R
\name{geos_is_empty}
\alias{geos_is_empty}
\alias{geos_is_simple}
\alias{geos_has_z}
\alias{geos_is_closed}
\title{Unary geometry predicates}
\usage{
geos_is_empty(x)

geos_is_simple(x)

geos_has_z(x)

geos_is_closed(x)
}
\arguments{
\item{x}{A geometry-like object}
}
\value{
A logical vector
}
\description{
Unary geometry predicates take one geometry and  return a logical vector.
}
\details{
\describe{
\item{\code{\link[=geos_is_empty]{geos_is_empty()}}}{
Returns \code{TRUE} if \code{x} contains no coordinates.
}
\item{\code{\link[=geos_is_simple]{geos_is_simple()}}}{
Returns \code{TRUE} if \code{x} does not intersect itself.
}
\item{\code{\link[=geos_has_z]{geos_has_z()}}}{
Returns \code{TRUE} if \code{x} contains z coordinates.
}
\item{\code{\link[=geos_is_closed]{geos_is_closed()}}}{
Returns \code{TRUE} if the first coordinate of \code{x} is also the
last coordinate of \code{x}.
}
}
}
\examples{
geos_is_empty(geo_wkt("POINT EMPTY"))
geos_is_empty(geo_wkt("POINT (30 10)"))



}
