% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-strtree.R
\name{geos_nearest}
\alias{geos_nearest}
\title{Find the closest feature}
\usage{
geos_nearest(geom, tree)
}
\arguments{
\item{geom}{A \link[=as_geos_geometry]{GEOS geometry vector}}

\item{tree}{A \code{\link[=geos_strtree]{geos_strtree()}}}
}
\value{
An integer vector of length \code{geom} containing the index
of \code{tree} that is closest to each feature in \code{geom}.
}
\description{
Finds the closest item index in \code{tree} to \code{geom}, vectorized along \code{geom}.
}
