% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-io.R
\name{geos_read_wkt}
\alias{geos_read_wkt}
\alias{geos_write_wkt}
\title{Read and write well-known text}
\usage{
geos_read_wkt(wkt)

geos_write_wkt(geom)
}
\arguments{
\item{wkt}{A vector of well-known text}

\item{geom}{A \link[=as_geos_geometry]{GEOS geometry vector}}
}
\description{
Read and write well-known text
}
\examples{
geos_read_wkt("POINT (30 10)")
geos_write_wkt(as_geos_geometry("POINT (30 10)"))

}
