% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-vector.R
\name{geos_segment_intersection}
\alias{geos_segment_intersection}
\title{Calculate the intersection of two segments}
\usage{
geos_segment_intersection(x, y)
}
\arguments{
\item{x, y}{Two \code{\link[=geo_segment]{geo_segment()}}s}
}
\value{
A \code{\link[=geo_xy]{geo_xy()}}
}
\description{
Calculate the intersection of two segments
}
\examples{
geos_segment_intersection(
  geo_segment(geo_xy(0, 0), geo_xy(10, 10)),
  geo_segment(geo_xy(10, 0), geo_xy(0, 10))
)

}
