% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-strtree.R
\name{geos_strtree}
\alias{geos_strtree}
\alias{as_geos_strtree}
\alias{as_geos_strtree.default}
\alias{as_geos_strtree.geos_strtree}
\alias{as_geos_strtree.geos_geometry}
\alias{geos_strtree_data}
\title{Create a GEOS STRTree}
\usage{
geos_strtree(geom)

as_geos_strtree(x, ...)

\method{as_geos_strtree}{default}(x, ...)

\method{as_geos_strtree}{geos_strtree}(x, ...)

\method{as_geos_strtree}{geos_geometry}(x, ...)

geos_strtree_data(tree)
}
\arguments{
\item{geom}{A \link[=as_geos_geometry]{GEOS geometry vector}}

\item{x}{An object to convert to a \code{\link[=geos_strtree]{geos_strtree()}}}

\item{...}{Unused}

\item{tree}{A \code{\link[=geos_strtree]{geos_strtree()}}}
}
\value{
A geos_str_tree object
}
\description{
Create a GEOS STRTree
}
