% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-coord.R
\name{is_geo_coord}
\alias{is_geo_coord}
\alias{geo_coord}
\alias{vec_ptype2.geo_coord}
\alias{vec_cast.geo_coord}
\title{Generics common to geo_coord_*() classes}
\usage{
is_geo_coord(x)

geo_coord()

\method{vec_ptype2}{geo_coord}(x, y, ...)

\method{vec_cast}{geo_coord}(x, to, ...)
}
\arguments{
\item{x}{A (possibly) \code{\link[=geo_coord_point]{geo_coord_point()}}, \code{\link[=geo_coord_multipoint]{geo_coord_multipoint()}},
\code{\link[=geo_coord_linestring]{geo_coord_linestring()}}, \code{\link[=geo_coord_multilinestring]{geo_coord_multilinestring()}},
\code{\link[=geo_coord_polygon]{geo_coord_polygon()}}, or \code{\link[=geo_coord_multipolygon]{geo_coord_multipolygon()}}, or \code{\link[=geo_xy]{geo_xy()}}}

\item{y, to}{See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}

\item{...}{Unused}
}
\description{
Generics common to geo_coord_*() classes
}
