% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-tbl.R
\name{is_geo_tbl}
\alias{is_geo_tbl}
\alias{geo_tbl}
\alias{vec_ptype2.geo_tbl}
\alias{vec_cast.geo_tbl}
\title{Generics common to geo_tbl_*() classes}
\usage{
is_geo_tbl(x)

geo_tbl()

\method{vec_ptype2}{geo_tbl}(x, y, ...)

\method{vec_cast}{geo_tbl}(x, to, ...)
}
\arguments{
\item{x}{A (possibly) \code{\link[=geo_tbl_point]{geo_tbl_point()}}, \code{\link[=geo_tbl_multipoint]{geo_tbl_multipoint()}},
\code{\link[=geo_tbl_linestring]{geo_tbl_linestring()}}, \code{\link[=geo_tbl_multilinestring]{geo_tbl_multilinestring()}},
\code{\link[=geo_tbl_polygon]{geo_tbl_polygon()}}, or \code{\link[=geo_tbl_multipolygon]{geo_tbl_multipolygon()}}, or \code{\link[=geo_xy]{geo_xy()}}}

\item{y, to}{See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}

\item{...}{Unused}
}
\description{
Generics common to geo_tbl_*() classes
}
