% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-coord-point.R
\name{new_geo_coord_point}
\alias{new_geo_coord_point}
\alias{new_geo_coord_multipoint}
\alias{validate_geo_coord_point}
\alias{validate_geo_coord_multipoint}
\alias{is_geo_coord_point}
\alias{is_geo_coord_multipoint}
\alias{format.geo_coord_point}
\alias{print.geo_coord_point}
\alias{format.geo_coord_multipoint}
\alias{print.geo_coord_multipoint}
\alias{vec_ptype_abbr.geo_coord_point}
\alias{vec_ptype_abbr.geo_coord_multipoint}
\alias{as_geo_coord_point}
\alias{as_geo_coord_point.default}
\alias{as_geo_coord_multipoint}
\alias{as_geo_coord_multipoint.default}
\title{S3 Details for (multi)point geometries}
\usage{
new_geo_coord_point(x = list(xy = geo_xy(), feature = integer(0)))

new_geo_coord_multipoint(x = list(xy = geo_xy(), feature = integer(0)))

validate_geo_coord_point(x)

validate_geo_coord_multipoint(x)

is_geo_coord_point(x)

is_geo_coord_multipoint(x)

\method{format}{geo_coord_point}(x, ...)

\method{print}{geo_coord_point}(x, ...)

\method{format}{geo_coord_multipoint}(x, ...)

\method{print}{geo_coord_multipoint}(x, ...)

\method{vec_ptype_abbr}{geo_coord_point}(x, ...)

\method{vec_ptype_abbr}{geo_coord_multipoint}(x, ...)

as_geo_coord_point(x, ...)

\method{as_geo_coord_point}{default}(x, ...)

as_geo_coord_multipoint(x, ...)

\method{as_geo_coord_multipoint}{default}(x, ...)
}
\arguments{
\item{x}{A (possibly) \code{\link[=geo_coord_point]{geo_coord_point()}} or \code{\link[=geo_coord_multipoint]{geo_coord_multipoint()}}}

\item{...}{Unused}
}
\description{
S3 Details for (multi)point geometries
}
