% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-coord-polygon.R
\name{new_geo_coord_polygon}
\alias{new_geo_coord_polygon}
\alias{new_geo_coord_multipolygon}
\alias{validate_geo_coord_polygon}
\alias{validate_geo_coord_multipolygon}
\alias{is_geo_coord_polygon}
\alias{is_geo_coord_multipolygon}
\alias{format.geo_coord_polygon}
\alias{print.geo_coord_polygon}
\alias{format.geo_coord_multipolygon}
\alias{print.geo_coord_multipolygon}
\alias{vec_ptype_abbr.geo_coord_polygon}
\alias{vec_ptype_abbr.geo_coord_multipolygon}
\alias{as_geo_coord_polygon}
\alias{as_geo_coord_polygon.default}
\alias{as_geo_coord_multipolygon}
\alias{as_geo_coord_multipolygon.default}
\title{S3 Details for (multi)polygon geometries}
\usage{
new_geo_coord_polygon(
  x = list(xy = geo_xy(), feature = integer(0), piece = integer(0))
)

new_geo_coord_multipolygon(
  x = list(xy = geo_xy(), feature = integer(0), part = integer(0), piece = integer(0))
)

validate_geo_coord_polygon(x)

validate_geo_coord_multipolygon(x)

is_geo_coord_polygon(x)

is_geo_coord_multipolygon(x)

\method{format}{geo_coord_polygon}(x, ...)

\method{print}{geo_coord_polygon}(x, ...)

\method{format}{geo_coord_multipolygon}(x, ...)

\method{print}{geo_coord_multipolygon}(x, ...)

\method{vec_ptype_abbr}{geo_coord_polygon}(x, ...)

\method{vec_ptype_abbr}{geo_coord_multipolygon}(x, ...)

as_geo_coord_polygon(x, ...)

\method{as_geo_coord_polygon}{default}(x, ...)

as_geo_coord_multipolygon(x, ...)

\method{as_geo_coord_multipolygon}{default}(x, ...)
}
\arguments{
\item{x}{A (possibly) \code{\link[=geo_coord_polygon]{geo_coord_polygon()}} or \code{\link[=geo_coord_multipolygon]{geo_coord_multipolygon()}}}

\item{...}{Unused}
}
\description{
S3 Details for (multi)polygon geometries
}
