% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-rect.R
\name{new_geo_rect}
\alias{new_geo_rect}
\alias{is_geo_rect}
\alias{validate_geo_rect}
\alias{vec_ptype_abbr.geo_rect}
\alias{format.geo_rect}
\alias{as_geo_rect}
\alias{as_geo_rect.default}
\alias{as_geo_rect.matrix}
\alias{as.matrix.geo_rect}
\alias{geo_size.geo_rect}
\title{S3 details for geo_rect}
\usage{
new_geo_rect(
  x = list(xmin = double(), ymin = double(), xmax = double(), ymax = double())
)

is_geo_rect(x)

validate_geo_rect(x)

\method{vec_ptype_abbr}{geo_rect}(x, ...)

\method{format}{geo_rect}(x, ...)

as_geo_rect(x, ...)

\method{as_geo_rect}{default}(x, ...)

\method{as_geo_rect}{matrix}(x, ...)

\method{as.matrix}{geo_rect}(x, ...)

\method{geo_size}{geo_rect}(x, ...)
}
\arguments{
\item{x}{A (possibly) \code{\link[=geo_rect]{geo_rect()}}}

\item{...}{Unused}
}
\description{
S3 details for geo_rect
}
