% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-segment.R
\name{new_geo_segment}
\alias{new_geo_segment}
\alias{is_geo_segment}
\alias{validate_geo_segment}
\alias{vec_ptype_abbr.geo_segment}
\alias{format.geo_segment}
\alias{as_geo_segment}
\alias{as_geo_segment.default}
\alias{geo_size.geo_segment}
\title{S3 details for geo_segment}
\usage{
new_geo_segment(x = list(start = geo_xy(), end = geo_xy()))

is_geo_segment(x)

validate_geo_segment(x)

\method{vec_ptype_abbr}{geo_segment}(x, ...)

\method{format}{geo_segment}(x, ...)

as_geo_segment(x, ...)

\method{as_geo_segment}{default}(x, ...)

\method{geo_size}{geo_segment}(x, ...)
}
\arguments{
\item{x}{A (possibly) \code{\link[=geo_segment]{geo_segment()}}}

\item{...}{Unused}
}
\description{
S3 details for geo_segment
}
