% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-tbl-point.R
\name{new_geo_tbl_point}
\alias{new_geo_tbl_point}
\alias{new_geo_tbl_multipoint}
\alias{validate_geo_tbl_point}
\alias{validate_geo_tbl_multipoint}
\alias{is_geo_tbl_point}
\alias{is_geo_tbl_multipoint}
\alias{format.geo_tbl_point}
\alias{print.geo_tbl_point}
\alias{format.geo_tbl_multipoint}
\alias{print.geo_tbl_multipoint}
\alias{vec_ptype_abbr.geo_tbl_point}
\alias{vec_ptype_abbr.geo_tbl_multipoint}
\alias{as_geo_tbl_point}
\title{S3 Details for (multi)point geometries}
\usage{
new_geo_tbl_point(x = list(xy = geo_xy(), feature = integer(0)))

new_geo_tbl_multipoint(x = list(xy = geo_xy(), feature = integer(0)))

validate_geo_tbl_point(x)

validate_geo_tbl_multipoint(x)

is_geo_tbl_point(x)

is_geo_tbl_multipoint(x)

\method{format}{geo_tbl_point}(x, ...)

\method{print}{geo_tbl_point}(x, ...)

\method{format}{geo_tbl_multipoint}(x, ...)

\method{print}{geo_tbl_multipoint}(x, ...)

\method{vec_ptype_abbr}{geo_tbl_point}(x, ...)

\method{vec_ptype_abbr}{geo_tbl_multipoint}(x, ...)

as_geo_tbl_point(x, ...)
}
\arguments{
\item{x}{A (possibly) \code{\link[=geo_tbl_point]{geo_tbl_point()}} or \code{\link[=geo_tbl_multipoint]{geo_tbl_multipoint()}}}

\item{...}{Unused}
}
\description{
S3 Details for (multi)point geometries
}
