% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-wkt.R
\name{new_geo_wkt}
\alias{new_geo_wkt}
\alias{is_geo_wkt}
\alias{validate_geo_wkt}
\alias{vec_ptype_abbr.geo_wkt}
\title{S3 details for geo_wkt}
\usage{
new_geo_wkt(x = character())

is_geo_wkt(x)

validate_geo_wkt(x)

\method{vec_ptype_abbr}{geo_wkt}(x, ...)
}
\arguments{
\item{x}{A character vector containing well-known text}

\item{...}{Unused}
}
\description{
S3 details for geo_wkt
}
\examples{
wkt <- geo_wkt("POINT (30 10)")
is_geo_wkt(wkt)

}
