% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-tbl.R
\name{unite_xy}
\alias{unite_xy}
\title{Create geometries from data frames}
\usage{
unite_xy(data, col, x, y, remove = TRUE)
}
\arguments{
\item{data}{A \code{\link[tibble:tibble]{tibble::tibble()}} or data frame}

\item{col}{A column name as a quoted string}

\item{x, y}{Column specifications for x and y coordinate
columns, respectively. Use \code{\link[dplyr:select]{dplyr::select()}} syntax.}

\item{remove}{Use \code{remove = FALSE} to keep \code{x} and \code{y} as columns
in the output.}
}
\value{
\code{data}, with a new column \code{col}
}
\description{
Create geometries from data frames
}
\examples{
tbl <- tibble(x = 1, y = 2)
unite_xy(tbl, "xy", x, y, remove = FALSE)

}
