% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-vctrs.R
\name{vctrs-methods}
\alias{vctrs-methods}
\alias{vec_cast.geos_geometry}
\alias{vec_ptype2.geos_geometry}
\title{Vctrs methods}
\usage{
vec_cast.geos_geometry(x, to, ...)

vec_ptype2.geos_geometry(x, y, ...)
}
\arguments{
\item{x, y, to, ...}{See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}.}
}
\description{
Vctrs methods
}
