% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-geometry.R, R/sf-compat.R, R/wk-compat.R
\name{as_geos_geometry}
\alias{as_geos_geometry}
\alias{as_geos_geometry.geos_geometry}
\alias{as_geos_geometry.character}
\alias{as_geos_geometry.blob}
\alias{as_geos_geometry.WKB}
\alias{as_geos_geometry.sfc}
\alias{as_geos_geometry.sf}
\alias{as_geos_geometry.wk_wkb}
\alias{as_geos_geometry.wk_wkt}
\alias{as_geos_geometry.wk_wksxp}
\title{Create GEOS Geometry Vectors}
\usage{
as_geos_geometry(x, ...)

\method{as_geos_geometry}{geos_geometry}(x, ...)

\method{as_geos_geometry}{character}(x, ...)

\method{as_geos_geometry}{blob}(x, ...)

\method{as_geos_geometry}{WKB}(x, ...)

\method{as_geos_geometry}{sfc}(x, ...)

\method{as_geos_geometry}{sf}(x, ...)

\method{as_geos_geometry}{wk_wkb}(x, ...)

\method{as_geos_geometry}{wk_wkt}(x, ...)

\method{as_geos_geometry}{wk_wksxp}(x, ...)
}
\arguments{
\item{x}{An object to be coerced to a geometry vector}

\item{...}{Unused}
}
\value{
A geos geometry vector
}
\description{
Create GEOS Geometry Vectors
}
\examples{
as_geos_geometry("LINESTRING (0 1, 3 9)")

}
