% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-sf.R, R/compat-wk.R, R/geos-geometry.R
\name{as_geos_geometry.sfc}
\alias{as_geos_geometry.sfc}
\alias{as_geos_geometry.sf}
\alias{as_geos_geometry.wk_wkb}
\alias{as_geos_geometry.wk_wkt}
\alias{as_geos_geometry.wk_xy}
\alias{as_geos_geometry.wk_xyz}
\alias{as_geos_geometry.wk_rct}
\alias{as_geos_geometry.wk_crc}
\alias{as_geos_geometry}
\alias{as_geos_geometry.geos_geometry}
\alias{as_geos_geometry.character}
\alias{as_geos_geometry.blob}
\alias{as_geos_geometry.WKB}
\alias{geos_geometry}
\title{Create GEOS Geometry Vectors}
\usage{
\method{as_geos_geometry}{sfc}(x, ...)

\method{as_geos_geometry}{sf}(x, ...)

\method{as_geos_geometry}{wk_wkb}(x, ...)

\method{as_geos_geometry}{wk_wkt}(x, ...)

\method{as_geos_geometry}{wk_xy}(x, ...)

\method{as_geos_geometry}{wk_xyz}(x, ...)

\method{as_geos_geometry}{wk_rct}(x, ...)

\method{as_geos_geometry}{wk_crc}(x, ...)

as_geos_geometry(x, ...)

\method{as_geos_geometry}{geos_geometry}(x, ...)

\method{as_geos_geometry}{character}(x, ..., crs = NULL)

\method{as_geos_geometry}{blob}(x, ..., crs = NULL)

\method{as_geos_geometry}{WKB}(x, ..., crs = NULL)

geos_geometry(crs = wk::wk_crs_inherit())
}
\arguments{
\item{x}{An object to be coerced to a geometry vector}

\item{...}{Unused}

\item{crs}{An object that can be interpreted as a CRS. See \code{\link[wk:wk_crs]{wk::wk_crs()}}.}
}
\value{
A geos geometry vector
}
\description{
Create GEOS Geometry Vectors
}
\examples{
as_geos_geometry("LINESTRING (0 1, 3 9)")

}
