% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-geometry.R
\name{as_geos_geometry}
\alias{as_geos_geometry}
\alias{as_geos_geometry.geos_geometry}
\alias{as_geos_geometry.character}
\title{Create GEOS Geometry Vectors}
\usage{
as_geos_geometry(x, ...)

\method{as_geos_geometry}{geos_geometry}(x, ...)

\method{as_geos_geometry}{character}(x, ...)
}
\arguments{
\item{x}{An object to be coerced to a geometry vector}

\item{...}{Unused}
}
\value{
A geos geometry vector
}
\description{
Create GEOS Geometry Vectors
}
\examples{
as_geos_geometry("LINESTRING (0 1, 3 9)")

}
