% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-wkb.R
\name{geo_wkb}
\alias{geo_wkb}
\title{Create and validate well-known binary}
\usage{
geo_wkb(x = list())
}
\arguments{
\item{x}{A \code{\link[=list]{list()}} of \code{\link[=raw]{raw()}} objects, each of which
represent well-known binary}
}
\value{
A \code{\link[=new_geo_wkb]{new_geo_wkb()}}
}
\description{
Like other geo types, \code{\link[=geo_wkb]{geo_wkb()}} doesn't convert its input
but does validate it using \code{\link[=validate_geo_wkb]{validate_geo_wkb()}}.
To skip validation, use \code{\link[=new_geo_wkb]{new_geo_wkb()}} with
the result of \code{vec_cast(list(...), .ptype = raw())}.
}
\examples{
# POINT (30 10) in WKB
wkb <- as.raw(
  c(
    0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
    0x00, 0x00, 0x00, 0x3e, 0x40, 0x00, 0x00, 0x00,
    0x00, 0x00, 0x00, 0x24, 0x40
  )
)
geo_wkb(list(wkb))

}
