% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-package.R
\docType{package}
\name{geos-package}
\alias{geos}
\alias{geos-package}
\title{geos: Open Source Geometry Engine (GEOS) R API}
\description{
Provides an R API to the Open Source Geometry Engine
  (GEOS) library and a vector format with which to efficiently
  store GEOS geometries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paleolimbot.github.io/geos}
  \item \url{https://github.com/paleolimbot/geos}
  \item Report bugs at \url{https://github.com/paleolimbot/geos/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
