% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-unary-atomic.R
\name{geos_area}
\alias{geos_area}
\title{Extract information from a GEOS geometry}
\usage{
geos_area(geom)
}
\arguments{
\item{geom}{A \link[=as_geos_geometry]{GEOS geometry vector}}
}
\value{
A vector of length \code{geom}
}
\description{
Extract information from a GEOS geometry
}
