% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-basic-strtree.R
\name{geos_basic_strtree}
\alias{geos_basic_strtree}
\alias{geos_basic_strtree_size}
\alias{geos_basic_strtree_finalized}
\alias{geos_basic_strtree_insert}
\alias{geos_basic_strtree_query}
\alias{geos_basic_strtree_query_filtered}
\title{Create a basic GEOS STRTree}
\usage{
geos_basic_strtree(node_capacity = 10L)

geos_basic_strtree_size(tree)

geos_basic_strtree_finalized(tree)

geos_basic_strtree_insert(tree, items)

geos_basic_strtree_query(tree, query, limit = NA, fill = FALSE)

geos_basic_strtree_query_filtered(
  tree,
  query,
  tree_geom,
  fun,
  ...,
  .chunk_size = 65536
)
}
\arguments{
\item{node_capacity}{The maximum number of child nodes that a node may have.
The minimum recommended capacity value is 4. If unsure, use a
default node capacity of 10.}

\item{tree}{A \code{\link[=geos_basic_strtree]{geos_basic_strtree()}}}

\item{items}{Items to add to the tree index}

\item{query}{Items with which to query the tree}

\item{limit}{The maximum number of matches in the tree to return}

\item{fill}{If \code{TRUE}, always returns \code{limit} matches per item in \code{query}
padded with \code{NA} if fewer than \code{limit} matches are found.}

\item{tree_geom}{A vctr coercible to \code{\link[=geos_geometry]{geos_geometry()}} whose indices
align with \code{tree}.}

\item{fun}{A vectorized binary predicate (e.g. \code{\link[=geos_intersects]{geos_intersects()}}) that
will be called with the tree geometry, the query geometry and any \code{...}
args passed.}

\item{...}{Passed to \code{fun}.}

\item{.chunk_size}{The approximate number of comparisons to pass to \code{fun}.}
}
\value{
A geos_basic_strtree object
}
\description{
An experimental alternative to the \code{\link[=geos_strtree]{geos_strtree()}} that provides a more
flexible interface and potentially faster spatial joins.
}
