% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-basic-strtree.R
\name{geos_basic_strtree}
\alias{geos_basic_strtree}
\alias{geos_basic_strtree_size}
\alias{geos_basic_strtree_finalized}
\title{Create a basic GEOS STRTree}
\usage{
geos_basic_strtree(node_capacity = 10L)

geos_basic_strtree_size(tree)

geos_basic_strtree_finalized(tree)
}
\arguments{
\item{node_capacity}{The maximum number of child nodes that a node may have.
The minimum recommended capacity value is 4. If unsure, use a
default node capacity of 10.}

\item{tree}{A \code{\link[=geos_basic_strtree]{geos_basic_strtree()}}}
}
\value{
A geos_basic_strtree object
}
\description{
Create a basic GEOS STRTree
}
