% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-topology.R
\name{geos_centroid}
\alias{geos_centroid}
\alias{geos_point_on_surface}
\alias{geos_node}
\alias{geos_boundary}
\alias{geos_envelope}
\alias{geos_convex_hull}
\alias{geos_minimum_rotated_rectangle}
\alias{geos_minimum_bounding_circle}
\alias{geos_minimum_width}
\alias{geos_minimum_clearance_line}
\alias{geos_minimum_clearance}
\title{Geometry summaries}
\usage{
geos_centroid(x, to = geo_ptype(x))

geos_point_on_surface(x, to = geo_ptype(x))

geos_node(x, to = geo_ptype(x))

geos_boundary(x, to = geo_ptype(x))

geos_envelope(x, to = geo_ptype(x))

geos_convex_hull(x, to = geo_ptype(x))

geos_minimum_rotated_rectangle(x, to = geo_ptype(x))

geos_minimum_bounding_circle(x, to = geo_ptype(x))

geos_minimum_width(x, to = geo_ptype(x))

geos_minimum_clearance_line(x, to = geo_ptype(x))

geos_minimum_clearance(x)
}
\arguments{
\item{x}{A geometry-like object}

\item{to}{A prototype created by \code{\link[=geo_ptype]{geo_ptype()}}}
}
\description{
These functions return geometries that summarize the input in
various ways.
}
\details{
Fully node a set of linestrings using the least possible number of nodes
while preserving all of the input nodes.
}
