% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-unary-geometry.R
\name{geos_centroid}
\alias{geos_centroid}
\alias{geos_boundary}
\alias{geos_minimum_width}
\alias{geos_minimum_clearance_line}
\alias{geos_minimum_rotated_rectangle}
\alias{geos_unary_union}
\alias{geos_point_on_surface}
\alias{geos_node}
\alias{geos_make_valid}
\alias{geos_unique_points}
\alias{geos_reverse}
\alias{geos_merge_lines}
\alias{geos_build_area}
\alias{geos_envelope}
\alias{geos_convex_hull}
\alias{geos_clone}
\title{Geometry transformers}
\usage{
geos_centroid(geom)

geos_boundary(geom)

geos_minimum_width(geom)

geos_minimum_clearance_line(geom)

geos_minimum_rotated_rectangle(geom)

geos_unary_union(geom)

geos_point_on_surface(geom)

geos_node(geom)

geos_make_valid(geom)

geos_unique_points(geom)

geos_reverse(geom)

geos_merge_lines(geom)

geos_build_area(geom)

geos_envelope(geom)

geos_convex_hull(geom)

geos_clone(geom)
}
\arguments{
\item{geom}{A \link[=as_geos_geometry]{GEOS geometry vector}}
}
\value{
A \link[=as_geos_geometry]{GEOS geometry vector} of length \code{geom}
}
\description{
Geometry transformers
}
