% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-vector.R
\name{geos_segment_intersection}
\alias{geos_segment_intersection}
\alias{geos_orientation_index}
\title{Segment operations}
\usage{
geos_segment_intersection(x, y)

geos_orientation_index(x, point)
}
\arguments{
\item{x, y}{\code{\link[=geo_segment]{geo_segment()}}s}

\item{point}{A \code{\link[=geo_xy]{geo_xy()}}}
}
\value{
\code{\link[=geos_segment_intersection]{geos_segment_intersection()}} returns a \code{\link[=geo_xy]{geo_xy()}};
\code{\link[=geos_orientation_index]{geos_orientation_index()}} returns -1, 0 or 1, depending if
the \code{point} lies to the right of (-1), is colinear with (0) or
lies to the left of (1) the segment (as judged from the start
of the segment looking towards the end).
}
\description{
Segment operations
}
\examples{
geos_segment_intersection(
  geo_segment(geo_xy(0, 0), geo_xy(10, 10)),
  geo_segment(geo_xy(10, 0), geo_xy(0, 10))
)

geos_orientation_index(
  geo_segment(geo_xy(0, 0), geo_xy(10, 10)),
  geo_xy(15, c(12, 15, 17))
)

}
