% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-version.R
\name{geos_version}
\alias{geos_version}
\alias{geos_capi_version}
\title{GEOS information}
\usage{
geos_version()

geos_capi_version()
}
\description{
GEOS information
}
\examples{
geos_version()
geos_capi_version()

}
