% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-xy.R
\name{new_geo_xy}
\alias{new_geo_xy}
\alias{is_geo_xy}
\alias{validate_geo_xy}
\alias{vec_ptype_abbr.geo_xy}
\alias{format.geo_xy}
\alias{as_geo_xy}
\alias{as_geo_xy.default}
\alias{as_geo_xy.matrix}
\alias{as.matrix.geo_xy}
\title{S3 details for geo_xy}
\usage{
new_geo_xy(x = list(x = double(), y = double()))

is_geo_xy(x)

validate_geo_xy(x)

\method{vec_ptype_abbr}{geo_xy}(x, ...)

\method{format}{geo_xy}(x, ...)

as_geo_xy(x, ...)

\method{as_geo_xy}{default}(x, ...)

\method{as_geo_xy}{matrix}(x, ...)

\method{as.matrix}{geo_xy}(x, ...)
}
\arguments{
\item{x}{A (possibly) \code{\link[=geo_xy]{geo_xy()}}}

\item{...}{Unused}
}
\description{
S3 details for geo_xy
}
